+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for multiple logs in herbicide quick form.
+- Add link to material type taxonomy.
 
 ### Changed